Fixed bug: Unable to delete file from the root dir

+++ ch.droptilllate.application/src/ch/droptilllate/application/controller/ViewController.java
@@ -225,7 +225,9 @@ public class ViewController {
 				}
 			}
 		}
+		if (folderList != null && folderList.size() > 0) {
 		database.deleteElement(folderList);
+		}
 		deleteTreeFiles(fileList);
 		deleteTreeFolders(folderList);
 		//CloseDatabase
